Add generated configuration and validation files for Huawei and Cisco devices
- Created expected configuration files for Huawei routers and switches (rtr01-ospf.huawei.cfg, sw01-access.huawei.cfg) based on Cisco source configurations.
- Added Cisco expected configuration for sw02-trunk (sw02-trunk.cisco.cfg) derived from Huawei source.
- Generated paste-ready configuration files for Huawei and Cisco devices to facilitate simulator validation.
- Implemented a validation worksheet for simulator testing, ensuring configurations load without syntax errors.
- Added tests for golden file consistency and correctness, ensuring renderer output matches expected configurations.
- Developed scripts for generating golden files and paste files, enhancing the testing and validation process.

+++ netmigrate/engine.py
@@ -98,7 +98,9 @@ def convert(
 
     lines = render(cfg)
 
-    duration_ms = int((time.perf_counter() - started) * 1000)
+    # No int() cast: a conversion takes well under a millisecond, so
+    # truncating to whole ms reports 0 for every file.
+    duration_ms = (time.perf_counter() - started) * 1000.0
 
     result = ConversionResult(
         source_vendor=source,
